Updated files
Finished the logic for random_pic_gen.js. Finished the html for the random pic in index.html. Added the picture oogieboogie.jpg to the gallery and optimized it. Updated gallery.html to include the new image. Changed the design of images in gallery.html to use figures and figcaptions. Figcaption will need to be updated once a summary of each picture is obtained.